Refactored talking_tree_storage until it can write an ogg file Currently, the ogg writer only supports writing game commands at the moment.
The format works as follows.

new_file() creates a file with an unixtimestamp.htogg

The serial number for the game stream is 0 on default.

VOIP would be 1.

The b_o_s page is currently a variant dictionary written in
the char * packet. The packet hold initalization variables as
written by write_header(const Dictionary & dict )

The commands occupy one page each. The whole command is written
in char * packet and the granulepos is the OS::get_ticks_usecs()
which is the timestamp since the engine started

The e_o_s packet is currently empty and has zero bytes.