This commit adds a basic voip replay.
The write and read somewhat works

The test project is in bomber2d_proto_ogg

There are ton of issues to be fixed.

I notice tons of static which is might be
due to a terrible record

Talkingtree must be refactored into a godot server

AudioCodec should be unifyied across both talkingtree and talking
tree storage.

I dont even dynamically create the voip header.

The implementation of player to audiostream is hacky

Audiostreamtalkingtree should be lockless

I should figure out a better logic for the packet reader because it
is terrible